move guid to hex id to new utils module
