+++ src/databricks/sql/backend/thrift_backend.py
@@ -15,8 +15,8 @@ from databricks.sql.backend.types import (
     SessionId,
     CommandId,
     BackendType,
-    guid_to_hex_id,
 )
+from databricks.sql.backend.utils import guid_to_hex_id
 
 try:
     import pyarrow
