+++ src/databricks/sql/backend/utils/__init__.py
@@ -1,0 +1,3 @@
+from .guid_utils import guid_to_hex_id
+
+__all__ = ["guid_to_hex_id"]
